+++ Chasm.Compatibility/Chasm.Compatibility.Attributes/CHANGELOG.md
@@ -1,11 +1,17 @@
 # Chasm.Compatibility.Attributes Changelog
 
 ### v2.1.0
-- 🧩 Added `net45` target. Now targets: `net8.0`, `net7.0`, `net6.0`, `net5.0`, `netcoreapp3.0`, `netcoreapp2.0`, `netcoreapp1.1`, `netcoreapp1.0`, `netstandard2.1`, `netstandard2.0`, `netstandard1.0`, `net45`, `net40`, `net35`;
+- 🧩 Added `netstandard1.1` and `net45` target. Now targets: `net8.0`, `net7.0`, `net6.0`, `net5.0`, `netcoreapp3.0`, `netcoreapp2.0`, `netcoreapp1.1`, `netcoreapp1.0`, `netstandard2.1`, `netstandard2.0`, `netstandard1.0`, `net45`, `net40`, `net35`;
 - ✨ Added `CallerArgumentExpressionAttribute` polyfill;
 - ✨ Added `CallerFilePathAttribute` polyfill;
 - ✨ Added `CallerLineNumberAttribute` polyfill;
 - ✨ Added `CallerMemberNameAttribute` polyfill;
+- ✨ Added `UnconditionalSuppressMessageAttribute` polyfill;
+- ✨ Added `CompilerFeatureRequiredAttribute` polyfill;
+- ✨ Added `InAttribute` polyfill;
+- ✨ Added `RequiresLocationAttribute` polyfill;
+- ✨ Added `RequiresPreviewFeaturesAttribute` polyfill;
+- ✨ Added `StackTraceHiddenAttribute` polyfill;
 
 ### v2.0.0
 - 🧩 Targets: `net8.0`, `net7.0`, `net6.0`, `net5.0`, `netcoreapp3.0`, `netcoreapp2.0`, `netcoreapp1.1`, `netcoreapp1.0`, `netstandard2.1`, `netstandard2.0`, `netstandard1.0`, `net40`, `net35`;